+++ src/App.js
@@ -4,15 +4,17 @@ import HeaderDetail from './components/templates/templates-header/HeaderDetail';
 import HeaderList from './components/templates/templates-header/HeaderList';
 import HeaderMain from './components/templates/templates-header/HeaderMain';
 import MainPage from './components/pages/pages-main/MainPage';
+import Detail from './components/templates/templates-detail/Detail';
 
 function App() {
   return (
     <>
-      <HeaderMain />
+      {/* <HeaderMain /> */}
       {/* <HeaderDetail /> */}
       {/* <HeaderList /> */}
       {/* <Reservation /> */}
-      <MainPage />
+      {/* <MainPage /> */}
+      <Detail />
     </>
   );
 }
