+++ src/components/UI/atoms/atoms-detail/ImageFrame.js
