+++ src/components/UI/atoms/atoms-detail/Score.js
