+++ src/components/UI/molecules/molecules-detail/BookingInfo.js
