+++ src/components/UI/molecules/molecules-detail/TypeInfo.js
